+++ examples/5qubit_gates.jl
@@ -7,12 +7,12 @@ function decompose_RX_on_5qubits()
     "num_qubits" => 5, 
     "depth" => 3,    
 
-    "elementary_gates" => ["H_1xCNot_23xI_4xI_5", "RX_2", "U3_4", "Identity"], 
-    "target_gate" => QCO.kron_single_qubit_gate(5, QCO.RXGate(π/4), "q4"),
+    "elementary_gates" => ["H_1xCNot_23xI_4xI_5", "RX_2", "CU3_34", "Identity"], 
+    "target_gate" => QCO.kron_two_qubit_gate(5, QCO.CRXGate(π/4), "q3", "q4"),
        
-    "U_θ_discretization" => [0, π/4],
-    "U_ϕ_discretization" => [0, -π/2],
-    "U_λ_discretization" => [0, π/2],    
+    "CU_θ_discretization" => [0, π/4],
+    "CU_ϕ_discretization" => [0, -π/2],
+    "CU_λ_discretization" => [0, π/2],    
 
     "RX_discretization" => [π/2],